requirements.txt
discord.py
google-api-python-client
google-auth
python-dotenv
